bf/requirements.txt
rospkg
catkin_pkg
pybitflyer
websocket-client
pandas
numpy
requests
TA-LIB
matplotlib
scikit-learn
xgboost
